Refactor grids into separate file
